show time from server run
