--- server/index.js
@@ -1,15 +1,0 @@
-const express = require('express');
-const app = express();
-const port = 8000;
-
-app.get('/', (request, response) => {
-    response.send('Hello Home!')
-});
-
-app.listen(port, (err) => {
-    if (err) {
-        return console.log('err', err);
-    }
-
-    console.log(`listening on ${port}`);
-});
